Get player stats from player
ATK, DEF, SPATK, SPDEF, and TYPE are now gotten from the user.  The
numeric values must sum to 125 and a valid type must be input in order
to continue.

Also tweaked the battle screen a bit.  Removed "MONSTER" before the
monster name, move SPATK and SPDEF to the right of ATK and DEF, and
removed the RATIO status.  Also moved the player section up a few lines.